Move ast_pool to ast directory
